requirements.txt
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.1
exceptiongroup==1.1.0
filelock==3.9.0
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.12.1
hyperframe==5.2.0
idna==2.10
iniconfig==2.0.0
joblib==1.2.0
numpy==1.24.2
opencv-python==4.7.0.68
packaging==23.0
pdfminer.six==20221105
pdfplumber==0.8.0
Pillow==9.4.0
pluggy==1.0.0
pycparser==2.21
pypdfium2==3.21.0
pytesseract==0.3.10
pytest==7.2.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
rfc3986==1.5.0
sacremoses==0.0.53
sentencepiece==0.1.97
six==1.16.0
sniffio==1.3.0
tokenizers==0.13.2
tomli==2.0.1
tqdm==4.64.1
transformers==4.26.1
typing_extensions==4.5.0
urllib3==1.26.14
Wand==0.6.11
